backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
google-generativeai==0.3.2
feedparser==6.0.11
pydantic==2.5.3
